requirements.txt
fastapi
uvicorn
python-multipart
pillow
httpx
python-dotenv
pytest